// src/api/models/new-artifact-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface NewArtifactTO
 */
export interface NewArtifactTO {
    /**
     * 
     * @type {string}
     * @memberof NewArtifactTO
     */
    name: any;
    /**
     * 
     * @type {string}
     * @memberof NewArtifactTO
     */
    description: any;
    /**
     * 
     * @type {string}
     * @memberof NewArtifactTO
     */
    fileType: any;
    /**
     * 
     * @type {string}
     * @memberof NewArtifactTO
     */
    svgPreview?: any;
}